Redesign of extension settings - part 1
It got cramped on single page, so a redesign was in order. This is only first part of redesign. Options were sorted in tabs and I added new option that allows installation or updating of npm modules with a single click. Also added tab that shows current IP for easier access to web player.