Redesign of extension settings - part 2
Second part of redesign. Now all the options have separate tabs. Added option to configure if label next to the remote button should be shown or not and an about page. Also to avoid problems that occurred when changing settings during streaming, window will now change when streaming is started or stopped.