Added server monitoring app
Server monitor is spawned from extension and runs separately in the background. App is responsible for running and restarting node server if it exited for some reason (e.g. killed by user or connection error). Previously this was not needed, as file-chooser was checking if server is running, but now with Nautilus integration this had to be added. Making this separate does not leave a zombie process when user restarts gnome-shell.