Fix slider.hasOwnProperty("setValue") not correctly detected
GNOME 3.34 uses a setter for slider value. Previous shell versions used setValue() function. We need to detect if slider has such function for backward compatibility, but using hasOwnProperty() to do so is wrong, because that function is in fact inherited from another class, thus it wont be detected with it.